poller/requirements.txt
psycopg2-binary==2.9.9
redis==5.0.1
influxdb-client==1.38.0
python-dotenv==1.0.0
